+++ src/DurationOption.php
@@ -41,9 +41,7 @@ class DurationOption extends AbstractAdminOption
         <script>
           jQuery(document).ready(function ($) {
 
-            var app = new Vue({
-
-              el: '#<?= $key; ?>',
+            var app = Vue.createApp({
 
               mounted: function () {
                 $('#<?= $key; ?> .option-wrap').fadeIn();
@@ -83,7 +81,7 @@ class DurationOption extends AbstractAdminOption
               methods: {
 
               }
-            });
+            }).mount('#<?= $key; ?>');
           });
         </script>
         <?php