+++ src/PostTypeSelectOption.php
@@ -120,7 +120,7 @@ class PostTypeSelectOption extends AbstractAdminOption
                     <p v-if="!items.length">
                         No <?= strtolower( $this->get_post_type_label('plural') ); ?> have been selected.
                     </p>
-                    <div v-for="item, i in items" class="item" :key="item">
+                    <div v-for="(item, i) in items" class="item" :key="item">
 
                         <span v-html="formatPostTitle(item, i)"></span>
 
@@ -151,9 +151,7 @@ class PostTypeSelectOption extends AbstractAdminOption
           console.log({key: <?= $key; ?>})
           jQuery(document).ready(function ($) {
 
-            var app = new Vue({
-
-              el: '#<?= $key; ?>-wrap',
+            var app = Vue.createApp({
 
               data: function () {
                 return {
@@ -241,7 +239,7 @@ class PostTypeSelectOption extends AbstractAdminOption
                   });
                 });
               }
-            });
+            }).mount('#<?= $key; ?>-wrap');
           })
         </script>
         <style>